+++ HISTORY.md
@@ -5,6 +5,7 @@ The file only contains the major change history. For full change history, please
 
 ## 0.4.0
 * Improvement (#49): Auto select the current date for date picker.
+* Improvement (#50): Allow user to customize empty sortable table.
 
 ## 0.3.9
 * Improvement (#2): Click sort arrow will show animation.